+++ osrsbox/testdata/all_items.json
