+++ osrsbox/testdata/all_monsters.json
