+++ osrsbox/testdata/all_prayers.json
