+++ osrsbox/testdata/bad_request_error.json
